// src/types/banner.ts
export interface Banner {
  bannerId: string
  targetId: string
  targetType: string
  targetName: string
  jumpUrl: string
  jumpMinAppId: string
  mediaType: number
  mediaUrl: string
  jumpType: number
  bannerLocation: string
  pageType: string
  clientType: number
  bgColor: string
  bannerTitle: string
  bannerDescription: string
  shareFlag: number
  sort: number
  forbiddenFlag: number
  insertTime: string
} 